adk-agent-demo/expense_assistant/requirements.txt
# Latest Google ADK version
google-adk>=1.10.0

# FastAPI and server components
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.6.0

# HTTP client for external APIs
requests>=2.31.0
httpx>=0.27.0

# Database components (PostgreSQL for sessions)
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25

# Azure services integration
azure-identity>=1.16.0
azure-search-documents>=11.4.0
azure-keyvault-secrets>=4.8.0

# Additional useful packages
python-multipart>=0.0.6
jinja2>=3.1.3

# Development and testing (optional)
pytest>=8.0.0
pytest-asyncio>=0.23.0